feat: started working on vap settings

+++ VencordAutoPatcher/Program.cs
@@ -20,48 +20,67 @@ internal class Program
         string installerPath = Path.Combine(vapDir, "VencordCLI.exe");
         string discordUpdateExe = Path.Combine(userProfile, "AppData", "Local", "Discord", "Update.exe");
 
-        if (!File.Exists(configPath))
+        if (args.Length < 1)
         {
-            Directory.CreateDirectory(vapDir);
-            ConsoleUtilities.Log("It looks like it's your first time running VAP");
-            ConsoleUtilities.Log("Do you want to setup Auto Start task? (y/n)");
-            string choice = ConsoleUtilities.Ask();
+            if (!File.Exists(configPath))
+            {
+                Directory.CreateDirectory(vapDir);
+                ConsoleUtilities.Log("It looks like it's your first time running VAP");
+                ConsoleUtilities.Log("Do you want to setup Auto Start task? (y/n)");
+                string choice = ConsoleUtilities.Ask();
 
-            if (choice == "y" || choice == "yes")
-                AutoRun.AddToStartup();
-            
-            File.WriteAllText(configPath, "");
+                if (choice == "y" || choice == "yes")
+                    AutoRun.AddToStartup();
+
+                File.WriteAllText(configPath, "");
+
+                ConsoleUtilities.Log("Press any key to close. Re-launch to open settings");
+                Console.ReadKey();
+                Environment.Exit(0);
+            }
         }
 
-        Console.Clear();
+        ConsoleUtilities.Log("Available options:");
+        ConsoleUtilities.Log("(r): clear firstlaunch flag");
+        string option = ConsoleUtilities.Ask();
+        if (option == "r") // TODO: make this a switch
+            File.Delete(configPath);
 
-        ConsoleUtilities.Log("Downloading Vencord CLI Installer");
-        using (WebClient web = new WebClient())
-            web.DownloadFile("https://github.com/Vencord/Installer/releases/latest/download/VencordInstallerCli.exe", installerPath);
 
-        ConsoleUtilities.Log("Successfully downloaded VencordInstallerCli.exe");
-
-        ConsoleUtilities.Log("Attempting to kill Discord process for patching.");
-        foreach (var process in Process.GetProcessesByName("Discord"))
-            try { process.Kill(); } catch { }
-
-        ConsoleUtilities.Log("Attempting to patch!");
-        Process patcher = new Process();
-        patcher.StartInfo.FileName = installerPath;
-        patcher.StartInfo.Arguments = "-install -branch auto";
-        patcher.Start();
-        patcher.WaitForExit();
-
-        ConsoleUtilities.Log("Patching completed.");
-
-        if (File.Exists(discordUpdateExe))
+        switch (args[0])
         {
-            ConsoleUtilities.Log("Relaunching Discord.");
-            Process.Start(discordUpdateExe);
-        }
-        else
-        {
-            ConsoleUtilities.Log("Update.exe not found, cannot restart Discord automatically!!!");
+            case "-autorun":
+                Console.Clear();
+
+                ConsoleUtilities.Log("Downloading Vencord CLI Installer");
+                using (WebClient web = new WebClient())
+                    web.DownloadFile("https://github.com/Vencord/Installer/releases/latest/download/VencordInstallerCli.exe", installerPath);
+
+                ConsoleUtilities.Log("Successfully downloaded VencordInstallerCli.exe");
+
+                ConsoleUtilities.Log("Attempting to kill Discord process for patching.");
+                foreach (var process in Process.GetProcessesByName("Discord"))
+                    try { process.Kill(); } catch { }
+
+                ConsoleUtilities.Log("Attempting to patch!");
+                Process patcher = new Process();
+                patcher.StartInfo.FileName = installerPath;
+                patcher.StartInfo.Arguments = "-install -branch auto";
+                patcher.Start();
+                patcher.WaitForExit();
+
+                ConsoleUtilities.Log("Patching completed.");
+
+                if (File.Exists(discordUpdateExe))
+                {
+                    ConsoleUtilities.Log("Relaunching Discord.");
+                    Process.Start(discordUpdateExe);
+                }
+                else
+                {
+                    ConsoleUtilities.Log("Update.exe not found, cannot restart Discord automatically!!!");
+                }
+                break;
         }
     }
 }
